requirements.txt
scrapy==2.5.0
ruamel.yaml==0.17.16